Remove duplicate function
